NAT: warning.
svn:sync-xref-src-repo-rev: r72611

+++ src/VBox/Devices/Network/slirp/debug.c
@@ -1,4 +1,4 @@
-/* $Id: debug.c 37741 2011-07-04 03:21:08Z noreply@oracle.com $ */
+/* $Id: debug.c 37742 2011-07-04 03:36:58Z noreply@oracle.com $ */
 /** @file
  * NAT - debug helpers.
  */
@@ -330,7 +330,7 @@ printTcpState(PFNRTSTROUTPUT pfnOutput, void *pvArgOutput,
                  void *pvUser)
 {
     size_t cb = 0;
-    const int idxTcpState = (int)pvValue;
+    const int idxTcpState = (int)(uintptr_t)pvValue;
     char *pszTcpStateName = (idxTcpState >= 0 && idxTcpState < TCP_NSTATES) ? g_apszTcpStates[idxTcpState] : "TCPS_INVALIDE_STATE";
     AssertReturn(RTStrCmp(pszType, "tcpstate") == 0, 0);
     cb += RTStrFormat(pfnOutput, pvArgOutput, NULL, 0, "%s", pszTcpStateName);
